+++ Expanded Cheat Menu/Changelog.md
@@ -6,6 +6,7 @@
 ## v11.3 (Unreleased)
 ### Changed:
 - Updated real time path marker to work for off-map colonists as well (ones in buildings).
+- EnumVars button to examine toolbar (only shows up when there's something to view).
 
 ### Fixed:
 - Set Colonists ages/etc didn't check if the colonist was valid (thanks tmpyemail).